implemented popup for deleting entry in HistoryFragment

+++ app/src/main/java/net/saidijamnig/healthapp/HistoryFragment.java
@@ -1,5 +1,7 @@
 package net.saidijamnig.healthapp;
 
+import android.app.AlertDialog;
+import android.graphics.drawable.ColorDrawable;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.LayoutInflater;
@@ -16,6 +18,7 @@ import net.saidijamnig.healthapp.adapter.HistoryListAdapter;
 import net.saidijamnig.healthapp.database.AppDatabase;
 import net.saidijamnig.healthapp.database.History;
 import net.saidijamnig.healthapp.database.HistoryDao;
+import net.saidijamnig.healthapp.databinding.EntryDeletePopupBinding;
 import net.saidijamnig.healthapp.databinding.FragmentHistoryBinding;
 
 import java.io.IOException;
@@ -31,6 +34,7 @@ public class HistoryFragment extends Fragment {
     private FragmentHistoryBinding binding;
     private ArrayList<History> historyElements = new ArrayList<>();
     HistoryListAdapter viewAdapter;
+    private int position;
 
     public HistoryFragment() {
         // Required empty public constructor
@@ -85,32 +89,56 @@ public class HistoryFragment extends Fragment {
 
     private void createLongPressListener() {
         viewAdapter.setOnItemLongClickListener((view, position) -> {
+            this.position = position;
             openDeleteRequestPopUp();
-            History historyElement = historyElements.get(position);
-            int uid = historyElement.uid;
+        });
+    }
 
-            String directory = requireActivity().getApplicationContext().getFilesDir().getAbsolutePath();
-            String imageName = historyElement.imageTrackName;
+    private void openDeleteRequestPopUp() {
+        AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(requireContext());
+        EntryDeletePopupBinding popupBinding = EntryDeletePopupBinding.inflate(LayoutInflater.from(requireContext()));
+        alertDialogBuilder.setView(popupBinding.getRoot());
+        alertDialogBuilder.setCancelable(false);
+
+        AlertDialog alertDialog = alertDialogBuilder.create();
+
+        popupBinding.buttonDeleteEntryNo.setOnClickListener(view1 -> alertDialog.dismiss());
+        popupBinding.buttonDeleteEntryYes.setOnClickListener(view1 -> {
+            deleteHistoryEntry();
+            alertDialog.dismiss();
+        });
+
+        if (alertDialog.getWindow() != null) {
+            alertDialog.getWindow().setBackgroundDrawable(new ColorDrawable(0));
+        }
+        alertDialog.show();
+    }
+
+    private void deleteHistoryEntry() {
+        History historyElement = historyElements.get(position);
+        int uid = historyElement.uid;
+
+        String directory = requireActivity().getApplicationContext().getFilesDir().getAbsolutePath();
+        String imageName = historyElement.imageTrackName;
+
+        if (imageName != null && !imageName.isEmpty()) {
             Path imagePath = Paths.get(directory, imageName);
-
             try {
                 Files.deleteIfExists(imagePath);
                 Log.i("TAG", "Successfully deleted image " + imageName);
             } catch (IOException e) {
                 Log.e("TAG", "Error deleting image file " + imageName + ": " + e);
-            } finally {
-                historyElements.remove(position);
-                viewAdapter.notifyItemRemoved(position);
-                deleteHistoryEntry(uid);
             }
-        });
+        } else {
+            Log.w("TAG", "Image name with id" + uid + " is null - Moving forward to removal of element deletion!");
+        }
+
+        historyElements.remove(position);
+        viewAdapter.notifyItemRemoved(position);
+        deleteHistoryEntryFromDatabase(uid);
     }
 
-    private void openDeleteRequestPopUp() {
-
-    }
-
-    private void deleteHistoryEntry(int uid) {
+    private void deleteHistoryEntryFromDatabase(int uid) {
         Thread thread = new Thread(() -> {
             historyDao.deleteHistoryEntryById(uid);
         });
